fix(StubQuoridorParser): moved to correct package
